btrfs-progs: remove the unused fs_info parameter for btrfs_csum_data()
The parameter @fs_info is not utilized at all, and there are already
several call sites passing NULL as @fs_info.

And there is no counter-part in kernel (we use crypto_shash_* interface
instead), there is no need to keep the parameter list the same.

So just remove the unused parameter.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>