btrfs-progs: restore: remove stale debug message
The message was introduced on 502d2872 ("btrfs-progs: restore: add
global verbose option"), and it was never changed. The debug message
shows the offset of one  EXTENT_DATA from the file being restore, but
it's only shown when it's not zero.

This field is non-zero when the extent was first snapshotted (or a
subvolume was created), and later changed partially, so the extent was
split in the part that remained the same and the one that changed.

It's not useful to have this message being printed without proper
context to the end user. The message itself isn't very useful in
debugging sessions, since it would be necessary more data about a
problematic file/extent. And given that the message is here for
more than 5 years being unnecessary printed, it's better to remove it
since it can annoy the users.

If a problem appears at this point in the code, a more appropriate debug
code could be introduced instead.

Link: https://github.com/kdave/btrfs-progs/issues/920
Reviewed-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>