btrfs-progs: remove btrfs_map_token structure
The structure is introduced by commit c979ffd78722 ("btrfs-progs: sync
accessors.[ch] from the kernel"), but no one is really utilizing it, as
in user space the whole extent buffer is just a contiguous user space
memory, no need to do any special page/folio related mapping.

Furthermore the kernel is also dropping the usage of btrfs_map_token,
replacing it with regular accessors.

So there is no need to keep the code and we can safely drop the
structure along with its helpers.

This time such removal does not only remove unused codes, but also
synchronise the codes with kernel.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>