btrfs-progs: docs: remove mkfs.btrfs note about memory usage
With the recent enhancements to --subvol and --inode-flags options, the
memory usage for them are just 4K+, and even for the older 8K+ memory
usage, it's really hard to consider them as heavy memory usage.

E.g. even if the end user specified over 1000 entries of
--subvol/--inode-flags, it only takes a little over 8MiB for the older
code or over 4MiB for the newer code.

Since we're in the user space and the year is 2025 not 1995, such memory
usage is far from heavy.

Just remove the paranoid note from the manual page.

Signed-off-by: Qu Wenruo <wqu@suse.com>