btrfs-progs: docs: update the compatibility about compression
There is a github issue that expressed confusion about the
"Compatibility" section of "ch-compression.rst".

The words in the man page is indeed confusing, and some points are no
longer correct either.

Considering how complex the direct IO and compression thing is, I didn't
come up with a correct answer until reading the code.

So update that section to provide a more straightforward result.

Issue: #1015
Signed-off-by: Qu Wenruo <wqu@suse.com>