btrfs-progs: mkfs: do zeroing before discard
mkfs.btrfs clears the very beginning and very end of a block device, to
prevent it from potentially being identified as another filesystem.

Change btrfs_prepare_device() so that this happens immediately before the
whole-device discard, rather than afterwards. This means that the contents
of these regions is either all zeroes (if discard did nothing) or indeterminate
(if discard was effective): either way it's clearing the other filesystem's
markers.

Pull-request: #1021
Signed-off-by: Mark Harmstone <mark@harmstone.com>
Signed-off-by: David Sterba <dsterba@suse.com>