btrfs-progs: fi resize: support offline/unmounted resize
Add support for resizing btrfs filesystems while unmounted using a new
option --offline, without requiring mount privileges or dealing with
mounted filesystem constraints.

Current limitations:

- increase size
- single-device filesystem

It works on both block devices and regular files. For regular files it
also truncates the file to the new size.

This provides an alternative for users who need to resize filesystems in
environments where mounting may not be possible or desirable, such as in
containers or during system recovery.

Pull-request: #1007
Signed-off-by: Leo Martins <loemra.dev@gmail.com>