btrfs-progs: count devices without SEARCH_TREE ioctl in get_fs_info()
To read the information about filesystem and devices we use the FS_INFO
ioctl, the num_devices however does not include the seeding device. To
compensate, search_chunk_tree_for_fs_info() enumerates the devices and
counts again using the SEARCH_TREE ioctl, which is privileged.

Due to that the whole get_fs_info() fails and prevents some commands to
run properly, eg. 'btrfs fi show' on a mounted path.

Change the counting to manual iteration up to the max_id which we have
read from the FS_INFO ioctl.

Signed-off-by: David Sterba <dsterba@suse.com>