btrfs-progs: device stats: add support for offline mode
Add new option --offline that reads device stats from the device (or a
file image) directly, ie. not using the GET_DEV_STATS ioctl. For that
the argument must be a block device or a file.

  $ btrfs device stats /tmp/img1
  ERROR: not a directory: /tmp/img1
  ERROR: to read device stats from a file image use --offline option

The output is the same as for online (mounted) case, with the exception
of missing devices that may not show full path in the output. Then the
device name is "devid:N" where N is the device id:

The following output is from 4 file backed images img[1234] that are
added to a filesystem as loop devices and then removed. The files exist
but the device scan cannot see them:

  $ btrfs device stats --offline /tmp/img1
  [/tmp/img1].write_io_errs    0
  [/tmp/img1].read_io_errs     0
  [/tmp/img1].flush_io_errs    0
  [/tmp/img1].corruption_errs  0
  [/tmp/img1].generation_errs  0
  [devid:2].write_io_errs    0
  [devid:2].read_io_errs     0
  [devid:2].flush_io_errs    0
  [devid:2].corruption_errs  0
  [devid:2].generation_errs  0
  [devid:3].write_io_errs    0
  [devid:3].read_io_errs     0
  [devid:3].flush_io_errs    0
  [devid:3].corruption_errs  0
  [devid:3].generation_errs  0
  [devid:4].write_io_errs    0
  [devid:4].read_io_errs     0
  [devid:4].flush_io_errs    0
  [devid:4].corruption_errs  0
  [devid:4].generation_errs  0

Signed-off-by: David Sterba <dsterba@suse.com>