btrfs-progs: tune: reset fs_info->new_chunk_tree_uuid at the end
The CodeQL scan points out a local variable is stored in a non-local
memory, in this case it's the uuid buffer for new chunk. There are error
paths leading to the label 'out' that do not reset it. This is not a
problem as this does not leak outside, but let's reset it in all cases.

Signed-off-by: David Sterba <dsterba@suse.com>