btrfs-progs: ci: limit codespell workflow only on release and requested branches
The Codespell workflow is useful and catches a lot of typos. However it
also makes development and CI use harder. Documentation updates do not
need the full CI so it's skipped. If a typo sneaks in any followup pull
request or devel update will fail.

We want to run the CI for code and updating and restarting the CI spends
all the free minutes available for the github actions.

At the slight inconvenience of missing some typos during development
limit the codespell workflow only on the 'release-test' branch or with
the prefix 'codespell/'.

[ci skip]

Signed-off-by: David Sterba <dsterba@suse.com>