btrfs-progs: check: add repair functionality for orphan dev extents
There is a bug report that btrfs-check --repair didn't repair an orphan
dev extent, which makes kernel to reject the fs due to the overly strict
chunk related checks.

However for btrfs-check, it's pretty easy to remove orphan device
extents (which do not have corresponding chunks), and update the device
item to reflect the used_bytes changes.

Add a repair function, btrfs_remove_dev_extent() which does the repair,
and call that function for both original and lowmem modes.

Pull-request: #1043
Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>