btrfs-progs: crypto: fix XXH3 build on AArch64
The XXH3 dispatcher exists only on x86_64, so use it conditionally. As
we want to print the implementation on other architectures as well add a
stub function replacing XXH_featureTest() when it does not exist. We can
reuse the value of XXH_VECTOR macro as it contains the used
implementation. For that it needs to be extracted from the actual XXH3
implementation maze. Example use of weak symbol and extern declaration.

Signed-off-by: David Sterba <dsterba@suse.com>