btrfs-progs: docs: add notes about interruption to scrub/dev-replace
Since the kernel patch is queued for v6.19, add notes to scrub and
dev-replace about the new behavior change.

For scrub it's not a huge deal as we can just resume using 'btrfs scrub
resume' command.
But for dev-replace one has to start from the beginning, thus it's
recommended to disable pm during dev-replace.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>