btrfs-progs: check: add repair ability for missing orphan root item
There is a known bug in older kernels that orphan items can be missing
for dropped subvolumes.

This makes the subvolumes unable to be removed on the next mount, and
recent kernel commit 4289b494ac55 ("btrfs: do not allow relocation of
partially dropped subvolumes") introduced one extra safe net to catch
such problem.

But unfortunately there is no way to repair it.

Add the repair ability to both the original and lowmem modes.

Reviewed-by: Boris Burkov <boris@bur.io>
Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>