btrfs-progs: do not dump leaf if the path is released inside __free_extent()
[BUG]
There is a bug report that btrfs-convert crashed during converting an
ext4 image which is almost full.

[CAUSE]
Just before the crash, btrfs-convert is committing the current
transaction but failed to locate the backref inside __free_extent().

Then it went through the error handling path, which prints an error
message and try to dump the leaf.

But in this particular case, the error code is not -ENOENT, thus the
path is already released, resuling path->nodes[0] to be NULL, and
btrfs_print_leaf() triggers a NULL pointer dereference.

[FIX]
The kernel version of btrfs_free_extent() is only dumping the tree for
-ENOENT error code. And patch "btrfs: refactor the error handling of
__btrfs_free_extent()" is submitted to make abort_and_dump() to only
dump the leaf if the path is not released.

So follow the same kernel patch, by only dumping the leaf if the path is
not released.

Issue: #1064
Signed-off-by: Qu Wenruo <wqu@suse.com>