btrfs-progs: ignore compile_commands.json
Clangd is the most commonly used LSP server for C, and it needs
a `compile_commands.json` file to get all the compiler flags for each
C file, and such file is normally the root marker to detect the root
directory of the project.

So it's pretty common to just put that file at the project top
directory.

Furthermore projects like the Linux kernel has already added that file
into gitignore, we should also ignore that file.

Signed-off-by: Qu Wenruo <wqu@suse.com>