btrfs-progs: corrupt-block: allow to specify the value for key corruption
I tried to use btrfs-corrupt-block -K to corrupt a INODE_REF type, but
unfortunately the field is always filled with random value, and
sometimes it even breaks the key order.

To make it more useful to reproduce the biflip recently reported, allow
"-K" option to work with "--value".

There are some minor points to note though:

- (u64)-1 will not work
  We use that value if detect if "--value" is specified.
  For most cases we do not have key objectid/offset set to (u64)-1, so
  it should not be a big deal.

- Will keep the random value behavior if --value is not specified
  So old behavior is still kept.

- Values over 255 will be truncated for key.type
  Just give a warning and do the usual truncation.

Signed-off-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>